requirements.txt
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.30.0,<0.31.0
requests>=2.32.0
lxml>=5.2.0
SQLAlchemy>=2.0.0
pydantic>=2.6.0,<3.0.0
PyYAML>=6.0.0
python-dateutil>=2.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
